Refactor and building of view class
